# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  query.GetOrderDto:
    properties:
      orderView:
        $ref: '#/definitions/query.OrderView'
        type: object
    type: object
...
  query.OrderView:
    properties:
      id:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: order context
  license: {}
  title: Order Application
  version: "1.0"
paths:
...
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Get order
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/query.GetOrderDto'
      summary: Get order
      tags:
      - order
...
swagger: "2.0"